# openapi/swagger-docs.yml
openapi: 3.0.3
info:
  title: MESS
  version: 1.0.0
  contact: {}
servers:
  - url: http://127.0.0.1
  - url: http://192.168.1.138
  - url: http://34.159.3.201
paths:
  /query/books:
    get:
      summary: RDF service - /query/books
      description: RDF service - /query/books
      operationId: rdfServiceQueryBooks
      parameters:
        - name: city
          in: query
          schema:
            type: string
            example: Iasi
      responses:
        '200':
          description: ''
...
  /query/books_by_ids:
    post:
      summary: RDF service - /query/books_by_ids
      description: RDF service - /query/books_by_ids
      operationId: rdfServiceQueryBooksByIds
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    example: '10333414'
                  example:
                    - '10333414'
                    - '10291544'
            examples:
              RDF service - /query/books_by_ids:
                value:
                  ids:
                    - '10333414'
                    - '10291544'
      responses:
        '200':
          description: ''
  /query/users:
    get:
      summary: RDF service - /query/users
      description: RDF service - /query/users
      operationId: rdfServiceQueryUsers
      parameters:
        - name: company
          in: query
          schema:
            type: string
            example: Lidl
        - name: city
          in: query
          schema:
            type: string
            example: Iasi
        - name: limit
          in: query
          schema:
            type: string
            example: '5'
      responses:
        '200':
          description: ''
...
  /recommendations/music:
    post:
      summary: Recommendation service - /recommendations/music
      description: Recommendation service - /recommendations/music
      operationId: recommendationServiceRecommendationsMusic
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                favorite_songs:
                  type: array
                  items:
                    type: object
                    properties:
                      album_name:
                        type: string
                        example: The Dark Leaves
                      artist_name:
                        type: string
                        example: Matt Pond PA
                      genres:
                        type: string
                        example: philly_indie stomp_and_holler
                      track_name:
                        type: string
                        example: Specks
                      uri:
                        type: string
                        example: spotify:track:4gAgdhTIx6T76baPfbXHQX
                  example:
                    - album_name: The Dark Leaves
                      artist_name: Matt Pond PA
                      genres: philly_indie stomp_and_holler
                      track_name: Specks
                      uri: spotify:track:4gAgdhTIx6T76baPfbXHQX
                    - album_name: Cup Chronicles
                      artist_name: Champagne Pa
                      genres: unknown
                      track_name: Diamonds From Africa
                      uri: spotify:track:4qGVbPTdfrvPGqeAKjry2N
            examples:
              Recommendation service - /recommendations/music:
                value:
                  favorite_songs:
                    - album_name: The Dark Leaves
                      artist_name: Matt Pond PA
                      genres: philly_indie stomp_and_holler
                      track_name: Specks
                      uri: spotify:track:4gAgdhTIx6T76baPfbXHQX
                    - album_name: Cup Chronicles
                      artist_name: Champagne Pa
                      genres: unknown
                      track_name: Diamonds From Africa
                      uri: spotify:track:4qGVbPTdfrvPGqeAKjry2N
      responses:
        '200':
          description: ''
...